multer added

+++ controllers/stuff.js
@@ -69,20 +69,34 @@ exports.modifyThing = (req, res, next) => {
     };
   
 exports.deleteThing = (req, res, next) => {
-    modelSauce.deleteOne({_id: req.params.id}).then(
-      () => {
-        res.status(200).json({
-          message: 'Deleted!'
+  modelSauce.findOne({_id: req.params.id}).then(
+    (sauce) => {
+      if (!sauce) {
+        return res.status(404).json({
+          error: new Error('No such thing!')
         });
       }
-    ).catch(
-      (error) => {
-        res.status(400).json({
-          error: error
+      if (sauce.userId !== req.auth.userId) {
+        return res.status(400).json({
+          error: new Error('Unauthorized request!')
         });
       }
-    );
-  };
+      modelSauce.deleteOne({_id: req.params.id}).then(
+        () => {
+          res.status(200).json({
+            message: 'Deleted!'
+          });
+        }
+      ).catch(
+        (error) => {
+          res.status(400).json({
+            error: error
+          });
+        }
+      );
+    } 
+  );
+};
 //Returns an array of all sauces in the database:
 exports.getAllStuff = (req, res, next) => {
     modelSauce.find().then(
